+++ src/mailman/handlers/tests/test_dmarc.py
@@ -408,6 +408,58 @@ Content-Disposition: inline
 
 """ + self._text)
 
+    def test_wrap_message_cc_with_reply_to(self):
+        self._mlist.dmarc_mitigate_action = DMARCMitigateAction.wrap_message
+        self._mlist.reply_goes_to_list = ReplyToMunging.point_to_list
+        msgdata = {'dmarc': True}
+        msg = mfs(self._text)
+        msg['Reply-To'] = 'user@example.com'
+        orig_msg = msg.as_string()
+        dmarc.process(self._mlist, msg, msgdata)
+        # We can't predict the Message-ID in the wrapper so delete it, but
+        # ensure we have one.
+        self.assertIsNotNone(msg.get('message-id'))
+        del msg['message-id']
+        self.assertMultiLineEqual(msg.as_string(), """\
+To: ant@example.com
+Subject: A subject
+X-Mailman-Version: X.Y
+Date: Fri, 1 Jan 2016 00:00:01 +0000
+MIME-Version: 1.0
+From: anne--- via Ant <ant@example.com>
+Cc: anne@example.com
+Reply-To: user@example.com
+Content-Type: message/rfc822
+Content-Disposition: inline
+
+""" + orig_msg)
+
+    def test_wrap_message_reply_to_with_cc(self):
+        self._mlist.dmarc_mitigate_action = DMARCMitigateAction.wrap_message
+        self._mlist.reply_goes_to_list = ReplyToMunging.no_munging
+        msgdata = {'dmarc': True}
+        msg = mfs(self._text)
+        msg['Cc'] = 'user@example.com'
+        orig_msg = msg.as_string()
+        dmarc.process(self._mlist, msg, msgdata)
+        # We can't predict the Message-ID in the wrapper so delete it, but
+        # ensure we have one.
+        self.assertIsNotNone(msg.get('message-id'))
+        del msg['message-id']
+        self.assertMultiLineEqual(msg.as_string(), """\
+To: ant@example.com
+Subject: A subject
+X-Mailman-Version: X.Y
+Date: Fri, 1 Jan 2016 00:00:01 +0000
+MIME-Version: 1.0
+From: anne--- via Ant <ant@example.com>
+Reply-To: anne@example.com
+Cc: user@example.com
+Content-Type: message/rfc822
+Content-Disposition: inline
+
+""" + orig_msg)
+
     def test_rfc2047_encoded_from(self):
         self._mlist.dmarc_mitigate_action = DMARCMitigateAction.munge_from
         msgdata = {'dmarc': True}